LPS-134118 Move files to components and hooks
